Add ImportComponent to handle importing FEN and PGN
* **App.jsx**
  - Remove existing input fields for FEN and PGN
  - Import and use the new `ImportComponent`

* **ChessBoard.jsx**
  - Remove `handleImport` function and related state variables

* **ImportComponent.jsx**
  - Create a new component with input fields for FEN and PGN
  - Add a button to trigger the import process
  - Pass the imported data to the necessary components

* **useChessStore.js**
  - Add actions to set and get FEN and PGN data
  - Ensure the imported data is stored and can be accessed by other components

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/w3bdesign/chess-tutor/pull/217?shareId=XXXX-XXXX-XXXX-XXXX).
